backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
httpx
gitpython
sentence-transformers
faiss-cpu
numpy
rank-bm25